fix flake8

+++ backend/users/views.py
@@ -31,6 +31,10 @@ class UserViewSet(UserViewSet):
             return Response(serializer.data, status=status.HTTP_201_CREATED)
 
         if request.method == 'DELETE':
+            if get_object_or_404(
+                Follow, user=user,
+                    author=author).DoesNotExist:
+                return Response(status=status.HTTP_204_NO_CONTENT)
             get_object_or_404(Follow, user=user, author=author).delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
